+++ src/process_train.py
@@ -1,7 +1,6 @@
 import argparse
-
 import multiprocessing
-
+from pathlib import Path
 
 import pandas as pd
 import spacy
@@ -9,13 +8,12 @@ from nltk import pos_tag
 
 from data import DATA_ROOT as DATA_PATH
 from evaluation import EVALUATION_ROOT as EVALUATION_PATH
-from .utils import get_logger, generate_uuid
 
+from .utils import generate_uuid, get_logger
 
 nlp = spacy.load("en_core_web_sm")
 num_cores = multiprocessing.cpu_count() - 1
 UUID = generate_uuid()
-logger = get_logger(run_id=UUID)
 
 
 def process_train_data(df, data_source, run_id, threshold=100, _logger=None):
@@ -32,39 +30,42 @@ def process_train_data(df, data_source, run_id, threshold=100, _logger=None):
         Path to the saved cleaned data. (Must change the name, sounds wrong)
 
     """
-    if not _logger:
-        _logger = logger
-    print("About to start processing train data.")
+    file_name = f"{DATA_PATH}/{run_id}_{data_source}_{threshold}_clean_train.csv"
+    if not Path(file_name).is_file():
+        if not _logger:
+            _logger = get_logger(run_id=UUID)
+        print("About to start processing train data.")
 
-    # remove columns below threshold
-    df_with_threshold = df.loc[:, (df.sum(axis=0) >= threshold)].copy()
+        # remove columns below threshold
+        df_with_threshold = df.loc[:, (df.sum(axis=0) >= threshold)].copy()
 
-    # create sum of rows
-    df_with_threshold["sum_of_features"] = df_with_threshold.sum(axis=1)
+        # create sum of rows
+        df_with_threshold["sum_of_features"] = df_with_threshold.sum(axis=1)
 
-    # drop words with no features
-    df_with_threshold.drop(
-        df_with_threshold[df_with_threshold.sum_of_features == 0].index, inplace=True
-    )
-    del df_with_threshold["sum_of_features"]
+        # drop words with no features
+        df_with_threshold.drop(
+            df_with_threshold[df_with_threshold.sum_of_features == 0].index,
+            inplace=True,
+        )
+        del df_with_threshold["sum_of_features"]
 
-    feature_metrics = df_with_threshold.sum(axis=0, skipna=True)
-    feature_metrics[1:].sort_values(ascending=False).to_csv(
-        f"{EVALUATION_PATH}/{run_id}_{data_source}_{threshold}_feature_metrics.csv",
-        header=["count_of_words_having_this_feature"],
-    )
+        feature_metrics = df_with_threshold.sum(axis=0, skipna=True)
+        feature_metrics[1:].sort_values(ascending=False).to_csv(
+            f"{EVALUATION_PATH}/{run_id}_{data_source}_{threshold}_feature_metrics.csv",
+            header=["count_of_words_having_this_feature"],
+        )
 
-    df_with_threshold.to_csv(
-        f"{DATA_PATH}/{run_id}_{data_source}_{threshold}_clean_train.csv"
-    )
+        df_with_threshold.to_csv(
+            f"{DATA_PATH}/{run_id}_{data_source}_{threshold}_clean_train.csv"
+        )
 
-    shape = df_with_threshold.shape
-    _logger.info(
-        f"Total number of words and hyponyms after applying the threshold of {threshold} is {shape}, \
-        where the first value is the number of words and the second the number of unique hyponyms\
-        i.e number of unique hyponyms that have more than {threshold} words"
-    )
-    return f"{DATA_PATH}/{run_id}_{data_source}_{threshold}_clean_train.csv"
+        shape = df_with_threshold.shape
+        _logger.info(
+            f"Total number of words and hyponyms after applying the threshold of {threshold} is {shape}, \
+            where the first value is the number of words and the second the number of unique hyponyms\
+            i.e number of unique hyponyms that have more than {threshold} words"
+        )
+    return file_name
 
 
 def _get_pos(word, is_spacy=True):
@@ -95,9 +96,10 @@ def enrich_data(clean_train, data_source, run_id):
     """
     threshold = clean_train.split("_")[-3]
     enriched_data = f"{DATA_PATH}/{run_id}_{data_source}_{threshold}_enriched_data.csv"
-    c_df = pd.read_csv(clean_train)
-    c_df["POS_TAG"] = c_df["actual_words"].apply(_get_pos)
-    c_df.to_csv(enriched_data, index=False)
+    if not Path(enriched_data).is_file():
+        c_df = pd.read_csv(clean_train)
+        c_df["POS_TAG"] = c_df["actual_words"].apply(_get_pos)
+        c_df.to_csv(enriched_data, index=False)
     return enriched_data
 
 
@@ -114,11 +116,12 @@ def process_enriched_data(enriched_data, data_source, run_id):
     """
     threshold = enriched_data.split("_")[-3]
     final_data = f"{DATA_PATH}/{run_id}_{data_source}_{threshold}_final_data.csv"
-    e_df = pd.read_csv(enriched_data)
-    dummy = pd.get_dummies(e_df["POS_TAG"], prefix="POS_TAG", drop_first=True)
-    del e_df["POS_TAG"]
-    e_df = pd.concat([e_df, dummy], axis=1)
-    e_df.to_csv(final_data, index=False)
+    if not Path(final_data).is_file():
+        e_df = pd.read_csv(enriched_data)
+        dummy = pd.get_dummies(e_df["POS_TAG"], prefix="POS_TAG", drop_first=True)
+        del e_df["POS_TAG"]
+        e_df = pd.concat([e_df, dummy], axis=1)
+        e_df.to_csv(final_data, index=False)
     return final_data
 
 
@@ -129,7 +132,7 @@ def main(script_args, _logger=None):
         _logger: A python logger object for logging run data
     """
     if not _logger:
-        _logger = logger
+        _logger = get_logger(run_id=UUID)
 
     if not script_args.run_id:
         script_args.run_id = UUID
@@ -137,16 +140,21 @@ def main(script_args, _logger=None):
     data_source = script_args.data_source
     print(script_args.data_source)
 
-    df = pd.read_csv(
-        f"{DATA_PATH}/{script_args.run_id}_word_{script_args.data_source}_train.csv"
-    )
-
-    df.set_index("actual_words", inplace=True)
-    clean_train = process_train_data(
-        df, script_args.data_source, script_args.run_id, script_args.threshold, _logger
-    )
-    enriched_data = enrich_data(clean_train, data_source, script_args.run_id)
-    process_enriched_data(enriched_data, data_source, script_args.run_id)
+    final_data = f"{DATA_PATH}/{script_args.run_id}_{script_args.data_source}_{script_args.threshold}_final_data.csv"
+    if not Path(final_data).is_file():
+        df = pd.read_csv(
+            f"{DATA_PATH}/{script_args.run_id}_word_{script_args.data_source}_train.csv"
+        )
+        df.set_index("actual_words", inplace=True)
+        clean_train = process_train_data(
+            df,
+            script_args.data_source,
+            script_args.run_id,
+            script_args.threshold,
+            _logger,
+        )
+        enriched_data = enrich_data(clean_train, data_source, script_args.run_id)
+        process_enriched_data(enriched_data, data_source, script_args.run_id)
 
     print("Done processing train data.")
